object id size increased to 4 bytes

+++ src/Game/Objects/IObject.h
@@ -17,7 +17,7 @@ class IObject {
   bool operator|(const IObject& other) const;
   bool operator&(const IObject& other) const;
 
-  uint16_t getId() const;
+  uint32_t getId() const;
   ObjectType getType() const;
   sf::Vector2f getPosition() const;
   sf::Vector2f getSize() const;
@@ -34,7 +34,7 @@ class IObject {
   virtual void destroy(Game& game, const ObjectRotation bulletRotation);
 
  protected:
-  uint16_t _id;
+  uint32_t _id;
   ObjectType _type;
 
   sf::Vector2f _position;
